# src/brick_game/server/swagger/model_error_message.go
/*
 * BrickGame
 *
 * BrickGame 3.0 REST API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type ErrorMessage struct {
	// Сообщение об ошибке
	Message string `json:"message,omitempty"`
}