# src/brick_game/server/swagger/model_game_info.go
/*
 * BrickGame
 *
 * BrickGame 3.0 REST API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type GameInfo struct {
	// Идентификатор игры
	Id int32 `json:"id,omitempty"`
	// Название игры
	Name string `json:"name,omitempty"`
}