# src/brick_game/server/swagger/model_game_state.go
/*
 * BrickGame
 *
 * BrickGame 3.0 REST API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type GameState struct {
	Width  int     `json:"width"`
	Height int     `json:"height"`
	Field  [][]int `json:"tiles"`
	Next   [][]int `json:"next"`
	// Опционально, если нужны
	Score     int  `json:"score,omitempty"`
	HighScore int  `json:"highScore,omitempty"`
	Level     int  `json:"level,omitempty"`
	Speed     int  `json:"speed,omitempty"`
	Pause     bool `json:"pause,omitempty"`
}