# src/brick_game/server/swagger/model_games_list.go
/*
 * BrickGame
 *
 * BrickGame 3.0 REST API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type GamesList struct {
	// Список игр
	Games []GameInfo `json:"games,omitempty"`
}