# src/brick_game/server/swagger/model_user_action.go
/*
 * BrickGame
 *
 * BrickGame 3.0 REST API
 *
 * API version: 1.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type UserAction struct {
	// Идентификатор действия
	ActionId int32 `json:"action_id,omitempty"`
	// Флаг, отвечающий за зажатие кнопки
	Hold bool `json:"hold,omitempty"`
}